@@ -1,8 +1,8 @@
 package com.malik.mytstunami.config;
 
 public class Constants {
-   public static final String URL ="http://192.168.2.248:8000";
-//   public static final String URL ="https://mytsunami.xyz";
+//   public static final String URL ="http://192.168.0.107:8000";
+   public static final String URL ="https://mytsunami.xyz";
 
    //Key Session
    public static final String KEY_USER_SESSION = "RSK123456";
